+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Period check test
 - POC PDFBox line wrapping
 - POC PDFBox special characters : U+2002 ('enspace')
 - Module pdfbox